literature/requirements.txt
google-genai>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.24.0
